feat: use cobra to provide subcommands, move sources to lib
