+++ lib/http/download.go
@@ -7,7 +7,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	fb "github.com/filebrowser/filebrowser"
+	fb "github.com/filebrowser/filebrowser/lib"
 	"github.com/hacdias/fileutils"
 	"github.com/mholt/archiver"
 )
