+++ lib/http/resource.go
@@ -13,7 +13,7 @@ import (
 	"strings"
 	"time"
 
-	fb "github.com/filebrowser/filebrowser"
+	fb "github.com/filebrowser/filebrowser/lib"
 	"github.com/hacdias/fileutils"
 )
