+++ lib/http/websockets.go
@@ -12,7 +12,7 @@ import (
 	"strings"
 	"time"
 
-	fb "github.com/filebrowser/filebrowser"
+	fb "github.com/filebrowser/filebrowser/lib"
 	"github.com/gorilla/websocket"
 )
