+++ lib/staticgen/staticgen.go
